requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
pillow
streamlit
mlflow
dvc